Started working on functions. header should have a static func return void. If a reflect function returns a value. Have that value stored as a local variable then set it in the static function in the header.

+++ Reflect/src/FileParser/FileParser.cpp
@@ -160,9 +160,9 @@ void FileParser::ReflectContainer(FileParsedData& fileData)
 	bool findStart = true;
 	ReflectContainerData& conatinerData = fileData.ReflectData.back();
 
-	int generatedBodyLine = fileData.Data.find(ReflectGeneratedBodykey, fileData.GeneratedBodyLineOffset);
+	int generatedBodyLine = static_cast<int>(fileData.Data.find(ReflectGeneratedBodykey, fileData.GeneratedBodyLineOffset));
 	assert(generatedBodyLine != -1 && "[FileParser::ReflectContainer] 'REFLECT_GENERATED_BODY()' is missing from a container.");
-	fileData.GeneratedBodyLineOffset = generatedBodyLine + strlen(ReflectGeneratedBodykey);
+	fileData.GeneratedBodyLineOffset = generatedBodyLine + static_cast<int>(strlen(ReflectGeneratedBodykey));
 	conatinerData.ReflectGenerateBodyLine = CountNumberOfSinceTop(fileData, generatedBodyLine, '\n') + 1;
 
 	while (true)
